Move notifications actions into separate package
