+++ reports-scheduler/src/main/kotlin/org/opensearch/reportsscheduler/ReportsSchedulerPlugin.kt
@@ -26,6 +26,7 @@
  */
 package org.opensearch.reportsscheduler
 
+import org.opensearch.OpenSearchStatusException
 import org.opensearch.jobscheduler.spi.JobSchedulerExtension
 import org.opensearch.jobscheduler.spi.ScheduledJobParser
 import org.opensearch.jobscheduler.spi.ScheduledJobRunner
@@ -73,10 +74,11 @@ import org.opensearch.env.Environment
 import org.opensearch.env.NodeEnvironment
 import org.opensearch.plugins.ActionPlugin
 import org.opensearch.plugins.Plugin
-import org.opensearch.reportsscheduler.action.NotificationsActions
+import org.opensearch.reportsscheduler.notifications.NotificationsActions
 import org.opensearch.repositories.RepositoriesService
 import org.opensearch.rest.RestController
 import org.opensearch.rest.RestHandler
+import org.opensearch.rest.RestStatus
 import org.opensearch.script.ScriptService
 import org.opensearch.threadpool.ThreadPool
 import org.opensearch.watcher.ResourceWatcherService
@@ -124,7 +126,12 @@ class ReportsSchedulerPlugin : Plugin(), ActionPlugin, JobSchedulerExtension {
         PluginSettings.addSettingsUpdateConsumer(clusterService)
         ReportDefinitionsIndex.initialize(client, clusterService)
         ReportInstancesIndex.initialize(client, clusterService)
-        NotificationsActions.initialize(client as NodeClient)
+        (client as? NodeClient)?.let { NotificationsActions.initialize(it) } ?: run {
+            throw OpenSearchStatusException(
+                "Unable to cast client to NodeClient for Notifications call",
+                RestStatus.INTERNAL_SERVER_ERROR
+            )
+        }
         return emptyList()
     }
 
@@ -184,17 +191,38 @@ class ReportsSchedulerPlugin : Plugin(), ActionPlugin, JobSchedulerExtension {
      */
     override fun getActions(): List<ActionPlugin.ActionHandler<out ActionRequest, out ActionResponse>> {
         return listOf(
-            ActionPlugin.ActionHandler(CreateReportDefinitionAction.ACTION_TYPE, CreateReportDefinitionAction::class.java),
-            ActionPlugin.ActionHandler(DeleteReportDefinitionAction.ACTION_TYPE, DeleteReportDefinitionAction::class.java),
-            ActionPlugin.ActionHandler(GetAllReportDefinitionsAction.ACTION_TYPE, GetAllReportDefinitionsAction::class.java),
-            ActionPlugin.ActionHandler(GetAllReportInstancesAction.ACTION_TYPE, GetAllReportInstancesAction::class.java),
+            ActionPlugin.ActionHandler(
+                CreateReportDefinitionAction.ACTION_TYPE,
+                CreateReportDefinitionAction::class.java
+            ),
+            ActionPlugin.ActionHandler(
+                DeleteReportDefinitionAction.ACTION_TYPE,
+                DeleteReportDefinitionAction::class.java
+            ),
+            ActionPlugin.ActionHandler(
+                GetAllReportDefinitionsAction.ACTION_TYPE,
+                GetAllReportDefinitionsAction::class.java
+            ),
+            ActionPlugin.ActionHandler(
+                GetAllReportInstancesAction.ACTION_TYPE,
+                GetAllReportInstancesAction::class.java
+            ),
             ActionPlugin.ActionHandler(GetReportDefinitionAction.ACTION_TYPE, GetReportDefinitionAction::class.java),
             ActionPlugin.ActionHandler(GetReportInstanceAction.ACTION_TYPE, GetReportInstanceAction::class.java),
-            ActionPlugin.ActionHandler(InContextReportCreateAction.ACTION_TYPE, InContextReportCreateAction::class.java),
+            ActionPlugin.ActionHandler(
+                InContextReportCreateAction.ACTION_TYPE,
+                InContextReportCreateAction::class.java
+            ),
             ActionPlugin.ActionHandler(OnDemandReportCreateAction.ACTION_TYPE, OnDemandReportCreateAction::class.java),
             ActionPlugin.ActionHandler(PollReportInstanceAction.ACTION_TYPE, PollReportInstanceAction::class.java),
-            ActionPlugin.ActionHandler(UpdateReportDefinitionAction.ACTION_TYPE, UpdateReportDefinitionAction::class.java),
-            ActionPlugin.ActionHandler(UpdateReportInstanceStatusAction.ACTION_TYPE, UpdateReportInstanceStatusAction::class.java)
+            ActionPlugin.ActionHandler(
+                UpdateReportDefinitionAction.ACTION_TYPE,
+                UpdateReportDefinitionAction::class.java
+            ),
+            ActionPlugin.ActionHandler(
+                UpdateReportInstanceStatusAction.ACTION_TYPE,
+                UpdateReportInstanceStatusAction::class.java
+            )
         )
     }
 }
